requirements.txt
Django==1.10.3
django-filter==1.0.1
django-money==0.10.1
djangorestframework==3.5.3
Markdown==2.6.8
psycopg2==2.6.2
py-moneyed==0.6.0
